requirements.txt
crewai
langachain_google_genai
load_dotenv